Implement sum dtype promotion rules

+++ cubed/tests/test_array.py
@@ -355,7 +355,7 @@ def test_default_spec_max_mem_exceeded():
 def test_reduction_multiple_rounds(tmp_path, executor):
     spec = xp.Spec(tmp_path, max_mem=110)
     a = xp.ones((100, 10), dtype=np.uint8, chunks=(1, 10), spec=spec)
-    b = xp.sum(a, axis=0)
+    b = xp.sum(a, axis=0, dtype=np.uint8)
     assert_array_equal(b.compute(executor=executor), np.ones((100, 10)).sum(axis=0))
